Summary of changes
Added the start playing game function to auth events file instead of
the game events file because its apart of the authorization process,
without being authorized to play you cant play the game board so I
decided to place it with the auth events file instead of the game
events file. I hope that was the right move, even though it doesn't
really intuitively feel like it is.