Summary of changes
Commented out .hide and .show functions to be able to see the entire
page all at once so that I don't have to complete the entire
authorization process before I can begin testing out the changes I make
to the game board.

Also changed the names to be more appropriate for the actions they are
invoking.