Summary of Changes
I switched the location for the gameBoard and startPlaying
functionality out of the auth files and into the game files.

I also uncommented the functionality for the .hide and .show
functionality to work again so that I can move forward with this step.

I also console logged some functions to make sure that they work.